Refactor Pomo DOne
