+++ src/components/Zen/ZenActions.vue
@@ -1,13 +1,6 @@
 <template>
   <div class="pomopause">
 
-    <v-badge v-if="!pomodoro.going && sink.itemsCount > 0" :content="sink.itemsCount" color="secondary">
-      <v-btn class='btn bg-primary pomo-btn pomo-box btn-main-start'
-        @click="sink.toggleSink()">
-        <v-icon class="icon" icon="mdi-faucet" />
-      </v-btn>
-    </v-badge>
-
     <v-btn class='btn bg-primary pomo-btn pomo-box font-press btn-main-start'
       v-if="!pomodoro.going && !pomodoro.settingUp"
       @click="emit('show-history')">
@@ -38,13 +31,9 @@
 </template>
 <script lang="ts" setup>
 import { usePomodoroStore } from "@/stores/pomodoro";
-import { useSinkStore } from "@/stores/sink";
 import { useSettingsStore } from "@/stores/settings";
-
 const pomodoro = usePomodoroStore();
 const settingsStore = useSettingsStore();
-const sink = useSinkStore();
-
 const emit = defineEmits<{
   (e: 'show-history'): void
 }>();
